requirements.txt
psutil
requests
platformdirs
windows_tools.antivirus